Fix handling of empty STK layers (question of why these exist remains) (#88)
It's a bit surprising that this would be the check that works, but sounds good. I'm guessing the empty STK layers were happening because we create a header without checking how many updates are available. If there are no updates, we should probably resize the file to remove the past header (or if it hasn't been written, then simply don't write the head to the file because the updates.size was 0)